+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - remove obsolete TypoScript files (#8)
 
 ### Fixed
+- Make the geocoding compatible with TYPO3 8LTS (#39)
 - Provide cli_dispatch.phpsh for 8.7 on Travis (#37)
 - Fix the DataMapper tests in TYPO3 8.7 (#32)
 - Require typo3-minimal for installing TYPO3 on Travis (#28)